modified on 2018-02-16

+++ src/main/java/com/jpa/repository/CustomerRepository.java
@@ -7,6 +7,6 @@ import com.jpa.entity.Customer;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called CustomerRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface CustomerRepository extends CrudRepository<Customer, Long> {
+public interface CustomerRepository extends CrudRepository<Customer, Integer> {
 
 }
