+++ src/main/java/com/jpa/repository/EmployeeRepository.java
@@ -7,6 +7,6 @@ import com.jpa.entity.Employee;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called CustomerRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface EmployeeRepository extends CrudRepository<Employee, Long> {
+public interface EmployeeRepository extends CrudRepository<Employee, Integer> {
 
 }
